rename utility to utils, silence some warnings
